requirement.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
celery-with-redis==3.0
certifi==2018.1.18
chardet==3.0.4
Django==1.8.2
django-celery==3.1.17
django-haystack==2.6.1
django-redis==4.9.0
django-redis-sessions==0.5.6
django-tinymce==2.6.0
fdfs-client-py==1.2.6
idna==2.6
itsdangerous==0.24
jieba==0.39
kombu==3.0.37
mutagen==1.40.0
Pillow==3.4.1
pkg-resources==0.0.0
pycryptodomex==3.5.1
PyMySQL==0.7.11
python-alipay-sdk==1.7.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
urllib3==1.22
uWSGI==2.0.15
Whoosh==2.7.4
